requirements.txt
# Deep Learning Framework
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# MedMNIST
medmnist==2.2.3

# Scientific Computing
numpy==1.24.3
scipy==1.11.3
pandas==2.0.3

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78
albumentations==1.3.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Model Utilities
timm==0.9.7                    # PyTorch Image Models (pre-trained models)
torchmetrics==1.2.0            # Metrics library
pytorch-lightning==2.1.0       # Optional: training framework

# Uncertainty Quantification
uncertainty-toolbox==0.1.1

# Experiment Tracking
wandb==0.15.12                 # Weights & Biases
tensorboard==2.14.1
mlflow==2.7.1

# Configuration
pyyaml==6.0.1
hydra-core==1.3.2              # Advanced config management
omegaconf==2.3.0

# Utilities
tqdm==4.66.1
scikit-learn==1.3.1
scikit-image==0.21.0

# Calibration
netcal==1.3.5                  # Network calibration

# Model Interpretation
captum==0.6.0                  # Model interpretability
grad-cam==1.4.8

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Jupyter
jupyter==1.0.0
ipywidgets==8.1.1
notebook==7.0.4